+++ jasperreports/src/net/sf/jasperreports/crosstabs/fill/calculation/BucketingService.java
@@ -150,7 +150,7 @@ public abstract class BucketingService
 		this.measureIndexes = new int[measureIndexList.size()];
 		for (int i = 0; i < measureIndexes.length; ++i)
 		{
-			measureIndexes[i] = measureIndexList.get(i).intValue();
+			measureIndexes[i] = measureIndexList.get(i);
 		}
 
 		this.retrieveTotal = retrieveTotal;
@@ -310,7 +310,7 @@ public abstract class BucketingService
 		}
 
 		measuresList.add(measure);
-		measureIndexList.add(Integer.valueOf(index));
+		measureIndexList.add(index);
 	}
