+++ jasperreports/src/net/sf/jasperreports/engine/data/ListOfArrayDataSource.java
@@ -59,9 +59,9 @@ public class ListOfArrayDataSource implements JRRewindableDataSource
 		
 		if (columnNames != null)
 		{
-			for(int i = 0; i < columnNames.length; i++)
+			for (int i = 0; i < columnNames.length; i++)
 			{
-				columnNamesMap.put(columnNames[i], Integer.valueOf(i));
+				columnNamesMap.put(columnNames[i], i);
 			}
 		}
 
@@ -101,7 +101,7 @@ public class ListOfArrayDataSource implements JRRewindableDataSource
 					new Object[]{jrField.getName()});
 		}
 
-		return currentRecord[fieldIndex.intValue()];
+		return currentRecord[fieldIndex];
 	}
